FSW-85 Fix C4389 by declaring or casting suspect variable to be of a more explicitly correct type
ViewDiagramm.cpp(5276,5612,5624,5642,5656,5668,5686,5700,5712,5731,5745,5757,5777,5791,5803,5817) (+1 squashed commits)
ViewTrend.cpp(2549,2620)
ViewStartup.cpp(500)
TrendDiagramm.cpp(3614,3728)
SoundPlayer.cpp(174,244,260,274,288,302,316)
InterfaceAcuLink_SERIAL.cpp(852)

+++ FabianHFO/InterfaceAcuLink_SERIAL.cpp
@@ -844,7 +844,7 @@ int CInterfaceAcuLink_SERIAL::sendAllMeasurementData()//CAcuLinkApp::BuildMsg()
 		{
 			dwBytesWRitten=m_comPort.Write(&(faSendBuf[iCount]),iRest);
 
-			if(dwBytesWRitten!=iRest)
+			if(dwBytesWRitten!=(DWORD)iRest)
 			{
 				CString sz;
 				sz.Format(_T("sendAcuLink2: %d"), dwBytesWRitten);
